Fix text overflow issue on en status card
Why:
Currently if the status text on the exposure notification status card on
the dashboard is too long the element layout is out of alignment.

This commit:
Refactors the exposure detection status card and adds flex sizing to
layout the text and the icon.
